+++ src/components/Avatar/styles.js
